Moves env_name out to its own files
